Move directory type test and add new case for auth options
